+++ src/Sequence/Primitive.php
@@ -54,7 +54,7 @@ final class Primitive implements Implementation
     }
 
     /**
-     * @return \Iterator<int, T>
+     * @return \Iterator<T>
      */
     public function iterator(): \Iterator
     {