+++ src/Set/Implementation.php
@@ -32,7 +32,7 @@ interface Implementation extends \Countable
     public function size(): int;
 
     /**
-     * @return \Iterator<int, T>
+     * @return \Iterator<T>
      */
     public function iterator(): \Iterator;